final changes from mitava

+++ product/models.py
@@ -5,7 +5,7 @@ from django.db import models
 class Category(models.Model):
     name = models.CharField(max_length=100, db_index=True)
     slug = models.SlugField(unique=True)
-    discount = models.IntegerField(default=10)
+    discount = models.DecimalField(max_digits=10, decimal_places=2, default=10.00)
 
     class Meta:
         ordering = ("name",)
